# specs/openapi.yaml
openapi: 3.0.0
info:
  title: Restaurant Ordering API
  description: A restaurant ordering system with authentication and restaurant management
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.restaurant.example.com/v1
    description: Production server
security:
  - BearerAuth: []
paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account with CUSTOMER role by default. Returns the created user object.
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
        "409":
          description: User already exists
  /auth/login:
    post:
      summary: Authenticate user and get JWT token
      description: Authenticates user credentials and returns a JWT token for accessing protected endpoints.
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid or missing request body
        "401":
          description: Invalid credentials
  /auth/logout:
    post:
      summary: Logout user (invalidate token)
      description: Invalidates the current JWT token, logging the user out of the system.
      operationId: logoutUser
      responses:
        "200":
          description: Logout successful
        "400":
          description: Bad request - malformed headers
        "401":
          description: Unauthorized - invalid or missing token
  /auth/me:
    get:
      summary: Get current user profile
      description: Returns the profile information for the currently authenticated user.
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request - malformed headers
        "401":
          description: Unauthorized - invalid or missing token
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for Spring Security authentication
  schemas:
    User:
      type: object
      description: User account information
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier (UUID)
        username:
          type: string
          description: Unique user login name
        email:
          type: string
          description: User email address
        firstName:
          type: string
          nullable: true
          description: User's first name
        lastName:
          type: string
          nullable: true
          description: User's last name
        roles:
          type: array
          description: User roles for authorization
          items:
            type: string
            enum: [CUSTOMER, RESTAURANT_OWNER, ADMIN]
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
    RegisterRequest:
      type: object
      description: Request body for user registration
      additionalProperties: false
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          description: Unique username for login (letters, numbers, underscores, hyphens only)
          pattern: "^[a-zA-Z0-9_-]+$"
          minLength: 3
          maxLength: 50
        email:
          type: string
          pattern: "^[a-zA-Z0-9_%+-]+(?:\\.[a-zA-Z0-9_%+-]+)*@[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]*[a-zA-Z0-9])?)*\\.[a-zA-Z]{2,63}$"
          maxLength: 120
          description: User's email address (standard format, max 120 chars)
        password:
          type: string
          pattern: "^[\\x20-\\x7E]+$"
          minLength: 8
          maxLength: 72
          description: ASCII printable characters only (space to tilde), max 72 bytes for bcrypt
        firstName:
          type: string
          nullable: true
          description: User's first name (optional, ASCII printable)
          pattern: "^[\\x20-\\x7E]*$"
          maxLength: 50
        lastName:
          type: string
          nullable: true
          description: User's last name (optional, ASCII printable)
          pattern: "^[\\x20-\\x7E]*$"
          maxLength: 50
    LoginRequest:
      type: object
      description: Request body for user authentication
      additionalProperties: false
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: User's registered username (letters, numbers, underscores, hyphens only)
          pattern: "^[a-zA-Z0-9_-]+$"
          minLength: 1
        password:
          type: string
          description: User's password (ASCII printable, at least one non-space)
          pattern: "^[\\x21-\\x7E][\\x20-\\x7E]*$"
          minLength: 1
    LoginResponse:
      type: object
      description: Response containing authentication token and user info
      properties:
        accessToken:
          type: string
          description: JWT access token
        tokenType:
          type: string
          description: Type of authentication token
          example: Bearer
        expiresIn:
          type: integer
          description: Token expiration time in seconds
        user:
          $ref: "#/components/schemas/User"
...